+++ src/pages/albums/components/SearchBar.jsx
@@ -1,63 +1,67 @@
-import React, { useEffect } from 'react'
+import React, { useEffect, useState } from 'react'
 
-export default function SearchBar() {
- 
-  async function fetchData(url) {
-    const rs = await fetch(url);
-    const json = await rs.json();
-    return json;
-  } 
+
+
+
+
+export default function SearchBar({searchValue, handleInputChange}) {
+
+  const [inputValue, setInputValue] = useState('');
   
-    useEffect(() => {
-      
-        fetchData(`https://jsonplaceholder.typicode.com/albums`)
-          .then((data) => console.log(data))
-      
-    }, [])
+
   
+  console.log(inputValue)
   
- 
-    return (
+
+
+
+  return (
     <form className="max-w-md mx-auto mt-4">
-  <label
-    htmlFor="default-search"
-    className="mb-2 text-sm font-medium text-gray-900 sr-only dark:text-white"
-  >
-    Search
-  </label>
-  <div className="relative">
-    <div className="absolute inset-y-0 start-0 flex items-center ps-3 pointer-events-none">
-      <svg
-        className="w-4 h-4 text-gray-500 dark:text-gray-400"
-        aria-hidden="true"
-        xmlns="http://www.w3.org/2000/svg"
-        fill="none"
-        viewBox="0 0 20 20"
+      <label
+        htmlFor="default-search"
+        className="mb-2 text-sm font-medium text-gray-900 sr-only dark:text-white"
       >
-        <path
-          stroke="currentColor"
-          strokeLinecap="round"
-          strokeLinejoin="round"
-          strokeWidth="2"
-          d="m19 19-4-4m0-7A7 7 0 1 1 1 8a7 7 0 0 1 14 0Z"
+        Search
+      </label>
+      <div className="relative">
+        <div className="absolute inset-y-0 start-0 flex items-center ps-3 pointer-events-none">
+          <svg
+            className="w-4 h-4 text-gray-500 dark:text-gray-400"
+            aria-hidden="true"
+            xmlns="http://www.w3.org/2000/svg"
+            fill="none"
+            viewBox="0 0 20 20"
+          >
+            <path
+              stroke="currentColor"
+              strokeLinecap="round"
+              strokeLinejoin="round"
+              strokeWidth="2"
+              d="m19 19-4-4m0-7A7 7 0 1 1 1 8a7 7 0 0 1 14 0Z"
+            />
+          </svg>
+        </div>
+        <input
+          type="search"
+          id="default-search"
+          className="block w-full p-4 ps-10 text-sm text-gray-900 border border-gray-300 rounded-lg bg-gray-50 focus:ring-blue-500 focus:border-blue-500 "
+          placeholder="Search Albums"
+          value={inputValue}
+          onChange={(e)=>{ 
+            setInputValue(e.target.value);
+            handleInputChange(e,inputValue)
+          }}
+          required
         />
-      </svg>
-    </div>
-    <input
-      type="search"
-      id="default-search"
-      className="block w-full p-4 ps-10 text-sm text-gray-900 border border-gray-300 rounded-lg bg-gray-50 focus:ring-blue-500 focus:border-blue-500 "
-      placeholder="Search Albums"
-      required
-    />
-    <button
-      type="submit"
-      className="text-white absolute end-2.5 bottom-2.5 bg-indigo-950 hover:bg-blue-800 focus:ring-4 focus:outline-none focus:ring-blue-300 font-medium rounded-lg text-sm px-4 py-2 "
-    >
-      Buscar
-    </button>
-  </div>
-</form>
+        <button
+          type="submit"
+          className="text-white absolute end-2.5 bottom-2.5 bg-indigo-950 hover:bg-blue-800 focus:ring-4 focus:outline-none focus:ring-blue-300 font-medium rounded-lg text-sm px-4 py-2 "
+
+        >
+          Buscar
+        </button>
+      </div>
+    </form>
 
   )
 }
